feat: Adiciona exception RecusoNaoEncontrado, e logica de service para proudto e carrinho

+++ src/main/java/com/pontodosjogos/pontodosjogos/controller/request/ItemCarrinhoRequest.java
@@ -2,11 +2,13 @@ package com.pontodosjogos.pontodosjogos.controller.request;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record ItemCarrinhoRequest(
         Long produtoId,
         Integer quantidade,
-        Double precoUnitarioNaCompra
+        BigDecimal precoUnitarioNaCompra
 ) {
 
 }
