+++ src/main/java/com/pontodosjogos/pontodosjogos/controller/response/ItemCarrinhoResponse.java
@@ -2,13 +2,15 @@ package com.pontodosjogos.pontodosjogos.controller.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record ItemCarrinhoResponse(
         Long id, // Para que o frontend possa remover/atualizar
         Long produtoId, // ID do produto
         String nomeProduto, // Nome para exibição
         Integer quantidade,
-        Double precoUnitarioNaCompra,
+        BigDecimal precoUnitarioNaCompra,
         Double subtotal
 ) {
